feat: Add Python 3.5 compatibility for shiv-built zipapps
We needed to be able to run a Shiv-built zipapp on a system that only
has Python 3.5 available. This likely won't be acceptable for upstream
inclusion in Shiv (the devs have previously stated that supporting 3.5
for built bundles but not for Shiv itself would be too confusing), but
the changes needed to support this are fairly minimal.